feat: merged hololive functions into one extension file
